pyproject.toml
[project]
name = "autoresearch-agents"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "openai>=1.0.0",
    "arxiv>=2.0.0",
    "requests>=2.30.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "numpy>=1.26.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.6.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.7.0",
    "streamlit>=1.55.0",

    # Azure
    "azure-core>=1.38.0",
    "azure-common>=1.1.28",
    "azure-search-documents>=11.6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "memory*", "tools*", "utils*"]